% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmeta.R
\name{abmeta}
\alias{abmeta}
\title{Analytical Bayesian Meta Analysis}
\usage{
abmeta(estimate, stderr, mu_prior = 0, se_prior = 1, interval = 0.9, RE = T)
}
\arguments{
\item{estimate}{A vector containing the effect-size}

\item{stderr}{A vector containing the standard error}

\item{mu_prior}{Prior for the mean}

\item{se_prior}{Prior for the se}

\item{interval}{Credibility intervals for the summary (default=0.9)}

\item{RE}{An argument indicating if RE or FE should be used (default RE=TRUE)}
}
\description{
Analytical Bayesian meta-analysis with random-effect (RE) or fixed-effect (FE).
}
