% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\docType{data}
\name{example1}
\alias{example1}
\title{This example dataset 1 for the meta-analysis}
\description{
This example dataset 1 for the meta-analysis
}
\keyword{error}
\keyword{estimates}
\keyword{standard}
