% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example4.R
\docType{data}
\name{example4}
\alias{example4}
\title{This example dataset 4 for the PPMN}
\description{
This example dataset 4 for the PPMN
}
\keyword{PPMN}
\keyword{and}
\keyword{errors}
\keyword{estimates}
\keyword{example}
\keyword{first}
\keyword{for}
