% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrabc.R
\name{extrabc}
\alias{extrabc}
\title{Extract ABC values}
\usage{
extrabc(
  obj,
  dist_threshold = NULL,
  warning = 50,
  interval = 0.9,
  n_dens = 100,
  line_position = 1.96,
  lab_size = 4,
  bin_limit = 6,
  bin_width = 0.5,
  xpos = 10,
  ypos_lim = 0.99,
  alpha_dens = 0.3
)
}
\arguments{
\item{obj}{An object from the abctoz function}

\item{dist_threshold}{A threshold to accept the number of iterations}

\item{interval}{Interval level}

\item{n_dens}{Number of density curves to generate}

\item{line_position}{the position of the vertical line in the histogram}

\item{bin_limit}{the limit to which the bins in the histogram extend}

\item{bin_width}{the width of the bins in the histogram}

\item{xpos}{Values for position of text along x-axis}

\item{ypos_lim}{A value in fraction of the maximum density that indicates limits of the y-axis}

\item{alpha_dens}{Transparency of the density lines}
}
\description{
Extract ABC values
}
