% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{A forest plot}
\usage{
forest(object, study_names = NULL, interval = 0.9, xlab = "Estimate")
}
\arguments{
\item{object}{An object generated by the meta function}

\item{study_names}{A character vector containing the names of each study}

\item{interval}{Credibility intervals for the summary (default=0.9)}

\item{xlab}{The name of the x-axis in the forest plot}
}
\description{
A forest plot that indicates all priors used in the model, all estimates included and the posterior distribution of the pooled
estimate.
}
