% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BF.R
\name{get_BF}
\alias{get_BF}
\title{Get Bayes Factor}
\usage{
get_BF(object, prior_names = NULL)
}
\arguments{
\item{prior_names}{names to place above the priors in the table}

\item{Object}{Model object from the meta function}
}
\description{
This simple function only works when one uses at least two priors
in the meta function. It returns a table with the number of simulation from each prior
or the fraction of the simulates. Dividing one by the other would then generate the
"Stochastic" Bayes factor. It is not a 'true' Bayes Factor, because normaly the prior odds
Are treated as fixed. Since there is no focus on hypothesis testing via Bayes Factors this is
not a major concern, but it gives a good posterior description of what prior model is
most plausible.
}
