% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmJAGS.R
\name{glmJAGS}
\alias{glmJAGS}
\title{Simple Generalized Linear Models with JAGS}
\usage{
glmJAGS(
  formula = NULL,
  data = NULL,
  family = "norm_ident",
  prior_mu = 0,
  prior_mu_se = 100,
  prior_dispersion = c(0.001, 0.001)
)
}
\arguments{
\item{formula}{Satandard notation y~x}

\item{data}{A dataset}

\item{family}{A set of "norm_ident", "norm_log", "gamma_ident", "gamma_log", "pois_ident", "pois_log", "negbinom_ident"
"negbinom_log"}

\item{prior_mu}{The prior for the mean of}

\item{prior_mu_se}{The prior for the standard error of the parameter}

\item{prior_dispersion}{The prior for the dispersion parameters (sigma or size) set via the gamma distribution}
}
\description{
Simple Generalized Linear Models with JAGS
}
\examples{
df  <- as.data.frame(array(rnorm(200), c(100,2)))
b0=2
b1=0.25
b2=0.1

df$V3 <- rpois(nrow(df), exp(b0+b1*df$V1+b2*df$V2))

results <- glmmJAGS(V3~V1+V2, data=df, family = "pois_log")

}
