% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop.R
\name{hop}
\alias{hop}
\title{Title}
\usage{
hop(
  object,
  group = NULL,
  predictor = NULL,
  link_function = NULL,
  nr_hops = 1500,
  resolution = 300,
  exponentiate = F,
  shift_b0 = 0,
  xlimit = c(-5, 10),
  ylimit = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{object}{An object generated by the mar function}

\item{group}{The group}

\item{predictor}{The predictor}

\item{link_function}{The link function}

\item{nr_hops}{The number of HOP-lines to be generated}

\item{resolution}{The resolution of the hop lines}

\item{exponentiate}{Exponentiation of the predictor gradient}

\item{shift_b0}{Shifting the intercept up or down by adding values}

\item{xlimit}{The limits of the x-axis}

\item{ylimit}{The limits of tye y-axis}

\item{xlab}{The x-label text}
}
\description{
Title
}
