% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar.R
\name{mar}
\alias{mar}
\title{Title}
\usage{
mar(
  estimate,
  stderr,
  parameter,
  predictor,
  link_function,
  grouping,
  gradient_dependency,
  prior_mu = 0,
  prior_mu_se = 1,
  prior_sd_max = 5,
  prior_weights = 1,
  get_prior_only = FALSE,
  lower_trunc = -1e+07,
  upper_trunc = 1e+07,
  n_chain = 2,
  n_thin = 1,
  n_seed = 666,
  n_iter = 10000,
  n_burnin = 1000,
  Rhat_warn = 1.01,
  Eff_warn = 1000
)
}
\arguments{
\item{estimate}{The model estimates}

\item{stderr}{The standard error of the model estimates}

\item{parameter}{The model parameter b0 or b1}

\item{predictor}{The name of the predictor}

\item{link_function}{The link of the model}

\item{grouping}{The names of the groups}

\item{gradient_dependency}{A value that expresses the gradient dependency, mean, variance, min, max, cv, etc.}

\item{prior_mu}{Priors for the mean}

\item{prior_mu_se}{Priors for the standard error of the mean of the mean}

\item{prior_sd_max}{Maximum value for the variance of the prior (it is a uniform prior starting at 0).}

\item{prior_weights}{The weights attached to each prior (prior odds)}

\item{get_prior_only}{Argument when true a list is returned for adjustment}

\item{lower_trunc}{lower truncation}

\item{upper_trunc}{upper truncation}

\item{n_chain}{The number of chains}

\item{n_thin}{Thinning step}

\item{n_seed}{Seed}

\item{n_iter}{Number of iterations}

\item{n_burnin}{Length of the burn-in}

\item{Rhat_warn}{When higher retursn an warning}

\item{Eff_warn}{When lower returns a warning}
}
\description{
Title
}
