% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_ppmn.R
\name{marginal_ppmn}
\alias{marginal_ppmn}
\title{Plots the marginal expectations of each edge-function with credibility intervals}
\usage{
marginal_ppmn(
  object,
  response,
  data,
  level = 0.9,
  nsim = 1000,
  n_grid = 100,
  ylim = NULL,
  ylab = NULL,
  plot_lab = F,
  lab_x = 0.6,
  lab_y = 0.85,
  round_lab = 1,
  lab_size = 3.5,
  lwd_exp = 0.8,
  lwd_col = "dodgerblue4",
  size_title = 8,
  size_text = 8,
  int_col = "dodgerblue3",
  int_alpha = 0.3,
  pt_alpha = 0.25
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{response}{The dependent (response) variable to display}

\item{data}{Data that contains x and y coordinates (points) to plot}

\item{level}{Credibility level (default level = 0.9)}

\item{nsim}{Number of simulations}

\item{n_grid}{Grid size of the added variable plot}

\item{ylim}{Limit of the y-axis}

\item{ylab}{Title of the y-axis}

\item{plot_lab}{Plot the performance of marginal function on the introduced data}

\item{lab_x}{Position of the label on the x-axis}

\item{lab_y}{Position of the label on the y-axis}

\item{round_lab}{Rounding the numbers in the label}

\item{lab_size}{Size of the label}

\item{lwd_exp}{Line width of the expected value}

\item{lwd_col}{Colour of the expected value}

\item{size_title}{Title size of the y-axis label}

\item{size_text}{Text size of the title}

\item{int_col}{Colour of the credibility intervals around the expected value}

\item{int_alpha}{Transparency of the credibility intervals around the expected value}

\item{pt_alpha}{Transparency of the points in the plot}
}
\description{
This is essentially an added variable plot. It displays the marginal relation of the
dependent variable and all independent variables for a single edge-function. It can also
display a label with performance metrics on the provided data. They only serve a check
after updating, and do not serve as actual metrics, the default is therefore FALSE. To
display only the line and intervals when no data is provided a dataset can be simulate of
the dependent and independent variables. The pt_alpha can consequently be set to 0 so the
points are not displayed.
}
