% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdplot.R
\name{pdplot}
\alias{pdplot}
\title{Posterior Density Plots}
\usage{
pdplot(
  object,
  interval = 0.9,
  display = "b1",
  order_predictor = NULL,
  order_group = NULL,
  top_label = 0.05,
  left_label = 0.1,
  title_size = 3,
  est_display = F,
  label_size = 1.6,
  line_width = 0.2,
  point_size = 0.5,
  err_bar_lwd = 0.2,
  xlab = "Parameter estimate",
  ylab = "Posterior density",
  xylab_size = 3,
  xtext_size = 6,
  xround = 1,
  xbreaks = 4
)
}
\arguments{
\item{object}{An object generated by the meta function}

\item{interval}{What credibility level to display}

\item{display}{Which model parameter to display b1 or b0 (currently only b1)}

\item{order_predictor}{The order of the predictors}

\item{order_group}{The order of the groups}

\item{top_label}{The space for the top label}

\item{left_label}{The space for the left labels}

\item{title_size}{Size of the titles}

\item{est_display}{Display the map, mean and se in the figures}

\item{label_size}{Size of the labels in the plots}

\item{line_width}{The line width}

\item{point_size}{The size of the point}

\item{err_bar_lwd}{The line width of the error bars}

\item{xlab}{The x-label text}

\item{ylab}{The y-label text}

\item{xylab_size}{Size of x and y label}

\item{xtext_size}{Size of the x-axis text}

\item{xbreaks}{Size of the breaks on the x-axis}

\item{scale_point_line}{Scale point and line as function of model parameter}

\item{xrounds}{rounding of the values on the x-axis}
}
\description{
This function displays the posterior density distribution of the chains. Thereby displaying to possible values of the model
parameters. The lower point indicates the Maximum APosterior Value (MAP), which is a 'fancy' way to say mode. The intervals indicate
the High Density Interval levels and colours red and blue indicate a negative or positive direction of the MAP, respectively.
}
\examples{
data("example1")

mod <-meta(estimate=example1$est, stderr=example1$se, RE=F,
          parameter=example1$parameter, predictor=example1$predictor,
          link_function=example1$link, grouping=example1$group)

fig1 <- pdplot(mod)

print(fig1$posterior_density$logit)

}
