% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpasz.R
\name{plotpasz}
\alias{plotpasz}
\title{Plot p-values as z-values}
\usage{
plotpasz(
  p,
  operator = NULL,
  line_position = 1.96,
  bin_limit = 6,
  bin_width = 0.5
)
}
\arguments{
\item{p}{a vector of numeric p-values}

\item{operator}{an operator indication < or >}

\item{line_position}{postion of the vertical line}

\item{bin_limit}{maximum value on the x-axis for the bins}

\item{bin_width}{width of the of the bins}
}
\description{
Plot p-values as z-values
}
