% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_comp.R
\name{pred_comp}
\alias{pred_comp}
\title{Predictive comparison}
\usage{
pred_comp(
  object,
  child = NULL,
  input_values,
  family = "normal",
  prior_mu1 = NULL,
  prior_sd1 = NULL,
  prior_mu0 = NULL,
  prior_sd0 = NULL,
  level = 0.9,
  nsim = 10000,
  xlim = NULL
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{child}{The child vertex of interest}

\item{input_values}{The input values at root vertices}

\item{family}{The appointed distributional family of the child vertex (default family = "normal") consisting of "normal", "gamma" or "beta"}

\item{prior_mu1}{Prior 1 for the mean of the the child vertex}

\item{prior_sd1}{SD 0 for the mean of the the child vertex}

\item{prior_mu0}{Prior 0 for the mean of the the child vertex}

\item{prior_sd0}{SD 0 for the mean of the the child vertex}

\item{level}{Credibility level (default level = 0.9)}

\item{nsim}{Number of simulations}

\item{xlim}{Limit of the x-axis within the plot}
}
\description{
Predictive comparison
}
