% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fit.R
\name{pred_fit}
\alias{pred_fit}
\title{Predictive fit function to assess performance}
\usage{
pred_fit(
  object,
  new_data,
  nsim = 1000,
  level = 0.9,
  rotate_x = 0,
  hjust = 0,
  seed = 123
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{new_data}{New dataset with all root nodes and least one child node present}

\item{nsim}{Number of simulations}

\item{level}{Credibility level (default level = 0.9)}

\item{rotate_x}{The angle to rotate the variable names on x-axis}

\item{hjust}{A number determining the height of the variable names on the x-axis}

\item{seed}{Seed value 123}
}
\description{
Predictive fit function to assess performance
}
