% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_edge.R
\name{predict_edge}
\alias{predict_edge}
\title{A generic function that defines and predicts with different functions}
\usage{
predict_edge(func, beta, trans, x)
}
\arguments{
\item{func}{Name of the function}

\item{beta}{Parameters of the functions}

\item{trans}{Transformation type applied over all independent variables}

\item{x}{matrix of independent variables}
}
\description{
A generic function that defines and predicts with different functions
}
