% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ppmn.R
\name{predict_ppmn}
\alias{predict_ppmn}
\title{Predict function for a PPMN object}
\usage{
predict_ppmn(object, new_data, nsim = 1000, print_drop = FALSE)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{new_data}{New data for prediction}

\item{nsim}{Number of simulations}

\item{print_drop}{Display the simulation progress (default print_drop = FALSE)}
}
\description{
Predict function for a PPMN object
}
