% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptoz.R
\name{ptoz}
\alias{ptoz}
\title{Title}
\usage{
ptoz(
  coef = NULL,
  sei = NULL,
  zi = NULL,
  pi = NULL,
  names = NULL,
  alpha = 0.05,
  absolute = F,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{coef}{The Effect-size}

\item{sei}{Standard error}

\item{zi}{z-value}

\item{pi}{p-value}

\item{names}{Additional names}

\item{alpha}{alpha level}
}
\description{
Title
}
