% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizep.R
\name{randomizep}
\alias{randomizep}
\title{Randomized p-values to transform to z-values}
\usage{
randomizep(p, operator, specific = T, seed = 123)
}
\arguments{
\item{p}{A vector of p-values from literature}

\item{operator}{A corresponding vector if this was "smaller" (e.g., p<.05) or "bigger" (p>.1)}

\item{specific}{If specific is TRUE all for p<.05 a random values is simulated as p~Uniform(0.01, 0.05) for <.01 Uniform(0.001, 0.01) etc.
If specific is False for every operator that is "smaller' p~Uniform(0, 0.05) often a proper approach.}

\item{seed}{Fixed seed set by default to 123}
}
\description{
Randomized p-values to transform to z-values
}
