% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_est.R
\name{root_est}
\alias{root_est}
\title{Root estimation function}
\usage{
root_est(
  object,
  child = NULL,
  target_value = NULL,
  prior = NULL,
  level = 0.9,
  plot_lab = F,
  lab_x = 0.75,
  lab_y = 0.75,
  round_lab = 1,
  lab_size = 3.5,
  kde_method = "Gaussian",
  kde_adjust = 1,
  kl_frac = 0.9
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{child}{The name of the child vertex observed in the network}

\item{target_value}{The value of the child vertex observed in the network}

\item{prior}{A data frame of value describing the prior values in the
root vertices}

\item{level}{Credibility level (default level = 0.9)}

\item{plot_lab}{Plots the point estimate (median) and standard error (se)
in the figure}

\item{lab_x}{Position of the label in percentage of the x-axis}

\item{lab_y}{Position of the label in percentage of the y-axis}

\item{round_lab}{Rounding the numbers in the label}

\item{lab_size}{Size of the label}

\item{kde_method}{The type of kernel density method (default kde_method = Gaussian)
can be Gaussian or Gamma}

\item{kde_adjust}{Adjustment value of the kernel bandwidth (default kde_adjust = 1)}

\item{kl_frac}{The strength of the update can be derived from the Kullenback-Leiber
divergence applied over the number of simulations (default kl_frac = 0.9) smaller
values mean less precision and broader intervals}
}
\description{
Root estimation function
}
