% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senscheck.R
\name{senscheck}
\alias{senscheck}
\title{Sensitivity check}
\usage{
senscheck(
  mod1,
  mod0,
  interval = 0.9,
  order_predictor = NULL,
  order_group = NULL,
  xlab = "Log(P(M1|Data, Info)/P(M0|Data, Info))"
)
}
\arguments{
\item{mod1}{First model preferable the model fitted using clearly defined priors}

\item{mod0}{An empty model only using a vague prior}

\item{interval}{What credibility level to display}

\item{order_predictor}{The order of the predictors}

\item{order_group}{The order of the groups}

\item{xlab}{The x-label text}
}
\description{
Sensitivity check
}
