% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ppmn.R
\name{update_ppmn}
\alias{update_ppmn}
\title{Update PPMN function}
\usage{
update_ppmn(
  object,
  new_data,
  nsim = 3000,
  level = 0.9,
  kl_frac = NULL,
  mad_constant = 1,
  seed = 123,
  max_lambda = 1000,
  covar = F,
  print_progress = F
)
}
\arguments{
\item{object}{A foundational or updated PPMN}

\item{new_data}{New dataset with all root nodes and least one child node present}

\item{nsim}{Number of simulations needed for updating}

\item{level}{Credibility level (default level = 0.9)}

\item{kl_frac}{The strength of the update can be derived from the Kullenback-Leiber divergence applied over the number of simulations (default kl_frac = NULL)}

\item{mad_constant}{the constant that scale the Median Absolute Deviation (default mda_constant = 1)}

\item{seed}{Seed value 123}

\item{max_lambda}{Maximum Lambda that can be set for Generalized Bayesian Updating (default max_lambda = 1000)}

\item{covar}{Covariance matrix that is calculated between model parameters (default covar = FALSE)}

\item{print_progress}{Progress report during updating}
}
\description{
Update PPMN function
}
