% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abctoz.R
\name{abctoz}
\alias{abctoz}
\title{Aproximated Bayesian Computation to derive Z (ABC-to-z)}
\usage{
abctoz(
  p,
  operator = NULL,
  nsim = 1e+05,
  prior_mu = c(1, 1),
  prior_sd = c(1, 2),
  prior_threshold = c(1.96, 0.1),
  prior_cens = c(3, 5),
  prior_uncens = c(1.25, 170),
  prior_odds_cens = 0.5,
  distribution = "z",
  density_steps = 200,
  seed = 1
)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{operator}{A vector that contains either <, > or NA (exact p-value)}

\item{nsim}{Number of simulations}

\item{prior_mu}{Prior values for mu(Z) following a truncated (0) normal distribution, which needs to notated c(mean, sd)}

\item{prior_sd}{Prior values for sd(Z) following a gamma distribution, which needs to be notated c(shape, rate)}

\item{prior_threshold}{Prior values for the threshold for 'significance'}

\item{prior_cens}{Prior values for the fractions of censored samples following a beta distribution, which needs to be notated as c(alpha, beta)}

\item{prior_uncens}{Prior values for the fractions of uncensored samples following a beta distribution, which needs to be notated as c(alpha, beta)}

\item{prior_odds_cens}{The likelihood of the prior odds}

\item{distribution}{Either z-distribution or shifted t-distribution with 1 df}

\item{density_steps}{An argument that determines the number of steps of the density function to derive the fit of the simulated to the data}

\item{seed}{Set the seed}
}
\description{
Aproximated Bayesian Computation to derive Z (ABC-to-z)
}
