% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abctoz.R
\name{abctoz}
\alias{abctoz}
\title{Aproximated Bayesian Computation to derive Z (ABC-to-z)}
\usage{
abctoz(
  z,
  nsim = 50000,
  prior_mu = c(0, 1),
  prior_sd = c(1, 1),
  prior_threshold = c(1.96, 0.1),
  prior_cens = c(1, 2),
  distribution = "t",
  print_progress = T,
  seed = 123
)
}
\arguments{
\item{z}{A numeric vector of z values}

\item{nsim}{Number of simulations}

\item{prior_mu}{Prior values for mu(Z) following a truncated (0) normal distribution, which needs to notated c(mean, sd)}

\item{prior_sd}{Prior values for sd(Z) following a gamma distribution, which needs to be notated c(shape, rate)}

\item{prior_cens}{Prior values for the fractions of censored samples following a beta distribution, which needs to be notated as c(alpha, beta)}

\item{print_progress}{Print the progress of simulations}

\item{seed}{Set the seed}

\item{z_cut}{Cut-off value for the z value}
}
\description{
Aproximated Bayesian Computation to derive Z (ABC-to-z)
}
