% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ppmn.R
\name{build_ppmn}
\alias{build_ppmn}
\title{Build the foundational Posterior Predictive Meta-analytic Network (PPMN)}
\usage{
build_ppmn(
  formula,
  data,
  txt_size = 3,
  vertex_size = 15,
  vertex_width = 0.21,
  vertex_height = 0.5,
  fun_width = 0.5,
  fun_height = 0.5,
  fun_lwd = 0.4,
  node_lwd = 0.4,
  arrow_size = 2,
  arrow_offset = 5.5,
  layout_type = "auto",
  circular = F
)
}
\arguments{
\item{formula}{An list of expressions that defines the network structure}

\item{data}{A list with estimates and standard errors to be synthesized}

\item{txt_size}{Size of the text with a visualization of the graph}

\item{vertex_size}{Diameter of the vertices in the DAG}

\item{vertex_width}{Width of the variable vertices in the Bipartite graph}

\item{vertex_height}{Height of the variable vertices in the Bipartite graph}

\item{fun_width}{Width of the vertices representing the edge-functions in the Bipartite graph}

\item{fun_height}{Height of the vertices representing the edge-functions in the Bipartite graph}

\item{fun_lwd}{Width of the edge-function vetices}

\item{node_lwd}{Width of the outlines variable vertices}

\item{arrow_size}{Size of the arrow lines}

\item{arrow_offset}{Distance of the arrows from the vertices}

\item{layout_type}{Type of layout}

\item{circular}{Circular layout or not (default circular = FALSE)}
}
\description{
Build the foundational Posterior Predictive Meta-analytic Network (PPMN)
}
