% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrabc.R
\name{extrabc}
\alias{extrabc}
\title{Extract ABC values}
\usage{
extrabc(
  obj,
  dist_threshold = 0.3,
  interval = 0.9,
  n_dens = 100,
  xpos = 6,
  alpha_dens = 0.3
)
}
\arguments{
\item{obj}{An object from the abctoz function}

\item{dist_threshold}{A threshold to accept the number of iterations}

\item{interval}{Interval level}

\item{n_dens}{Number of density curves to generate}

\item{xpos}{Values for position of text along x-axis}

\item{alpha_dens}{Transparency of the density lines}
}
\description{
Extract ABC values
}
