% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop.R
\name{hop}
\alias{hop}
\title{Hypothetical outcome plots}
\usage{
hop(
  object,
  group = NULL,
  predictor = NULL,
  link_function = NULL,
  nr_hops = 1500,
  hop_lwd = 1.2,
  hop_alpha = 0.05,
  hop_col = "dodgerblue1",
  expected_lwd = 0.8,
  expected_col = "dodgerblue4",
  hop_resolution = 300,
  pdp_resolution = 30,
  exp_disp = F,
  exp_axis = F,
  shift_b0 = 0,
  gradient_title = "MAP",
  gradient_col = c("magenta", "yellow", "cyan"),
  xlimit = c(-5, 10),
  ylimit = NULL,
  xlab = NULL,
  ylab = NULL,
  round_x_axis = 2,
  round_y_axis = 2,
  breaks_axis = 5,
  xlabsize = 8,
  ylabsize = 8,
  xtextsize = 10,
  ytextsize = 10
)
}
\arguments{
\item{object}{An object generated by the mar function}

\item{group}{The group}

\item{predictor}{The predictor}

\item{link_function}{The link function}

\item{nr_hops}{The number of HOP-lines to be generated}

\item{hop_lwd}{The line width of the hop-lines}

\item{hop_alpha}{The transparency of the hop-lines}

\item{hop_col}{The color (or Colour for the grammar freak) of the hop-lines}

\item{expected_lwd}{The line width of the hop-lines}

\item{expected_col}{The color of the line describing the expected value}

\item{hop_resolution}{The resolution of the hop lines}

\item{pdp_resolution}{The resolution of the pdp grid}

\item{exp_disp}{Exponentiation of the predictor gradient}

\item{exp_axis}{Take the exponent of the axis number}

\item{shift_b0}{Shifting the intercept up or down by adding values}

\item{gradient_title}{Title of the gradient in a pdp-plot}

\item{xlimit}{The limits of the x-axis}

\item{ylimit}{The limits of the y-axis}

\item{xlab}{The x-label text}

\item{ylab}{The y-label text}

\item{breaks_axis}{The number of breaks on the axis}

\item{xlabsize}{The size of the x-label text}

\item{ylabsize}{The size of the y-label text}

\item{xtextsize}{The size of the x-axis text}

\item{ytextsize}{The size of the y-axis text}

\item{gadient_col}{The color of the line describing the gradient}

\item{round_axis}{Round the number on the axis}
}
\description{
Hypothetical outcome plots
}
