% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpasz.R
\name{plotpasz}
\alias{plotpasz}
\title{Plot p-values as z-values}
\usage{
plotpasz(
  p,
  operator = NULL,
  line_position = 1.96,
  bin_limit = 6,
  bin_width = 0.5
)
}
\arguments{
\item{bin_width}{}
}
\description{
Plot p-values as z-values
}
