% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescheck.R
\name{rescheck}
\alias{rescheck}
\title{Residual check}
\usage{
rescheck(
  object,
  order_predictor = NULL,
  order_group = NULL,
  xlimit = 0.975,
  ylimit = c(0.025, 0.975),
  b0_out = T
)
}
\arguments{
\item{object}{An object generated by the meta function}

\item{order_predictor}{The order of the predictors}

\item{order_group}{The order of the groups}

\item{xlimit}{the upper quantile limit of 1/se or 1/n}

\item{ylimit}{The quantile limits residuals displayed in the figures}

\item{b0_out}{Display the intercept in the residuals, by default True and not displayed.}
}
\description{
This function can be used to asses the bias of the model output. It is a visual ('eye balling') check
for the distribution of the residuals. The residuals should be approximately 'randomly' distributed around 0 (origin) at the y-axis
relative to the inverse of the sample size (1/n) or inverse of the standard error (1/se). It is akin to classical funnel plots
If bias is clear a diagonal pattern would become visible the dashed red line would than not fall directly over the blue line.
}
